2026-05-21 worker filter

+++ app/Http/Controllers/WorkerController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Filters\Var1\WorkerFilter;
 use App\Http\Requests\Worker\IndexRequest;
 use App\Http\Requests\Worker\StoreRequest;
+use App\Http\Requests\Worker\UpdateRequest;
 use http\QueryString;
 use Illuminate\Http\Request;
 use App\Models\Worker;
@@ -13,41 +15,16 @@ class WorkerController extends Controller
 
     public function index(IndexRequest $request)
     {
-        $data = $request->validated();
-        $workerQuery = Worker::query();
+        $data = $request->validated(); // Данные запроса отвалидированные
 
-        if (isset($data['name'])) {
-            $workerQuery->where('name', 'like', "%{$data['name']}%");
-        }
-
-        if (isset($data['surname'])) {
-            $workerQuery->where('surname', 'like', "%{$data['surname']}%");
-        }
-
-        if (isset($data['email'])) {
-            $workerQuery->where('email', 'like', "%{$data['email']}%");
-        }
-
-        if (isset($data['from'])) {
-            $workerQuery->where('age', '>', $data['from']);
-        }
-
-        if (isset($data['to'])) {
-            $workerQuery->where('age', '<', $data['to']);
-        }
-
-        if (isset($data['description'])) {
-            $workerQuery->where('description', 'like', "%{$data['description']}%");
-        }
-
-        if (isset($data['is_married'])) {
-            $workerQuery->where('is_married', true);
-        }
-
-
-
-
+        $workerQuery = Worker::query(); // Создал билдер
+        $filter = new WorkerFilter($data);  // Это параметры фильтра
+        $filter->applyFilter($workerQuery); // Создаю фильтр и закидываю туда Билдер
+//        dd($bill->get()->toArray()); // Модифицированный запрос, но мы тут используем пагинатор.
         $workers = $workerQuery->paginate(4);
+
+
+
 
         return view('worker.index', compact('workers'));
 
@@ -69,42 +46,51 @@ class WorkerController extends Controller
 
     public function store(StoreRequest $request)
     {
+        $this->authorize('create', Worker::class);
+
         $data = $request->validated();
         $data['is_married'] = isset($data['is_married']);
+        $data['position_id'] = 1;
 
         Worker::create($data);
 
-        return redirect()->route('worker.index');
+        return redirect()->route('workers.index');
 
     }
 
 
     public function edit(Worker $worker)
     {
+        $this->authorize('update', $worker);
+
         return view('worker.edit', compact('worker'));
 
     }
 
-    public function update(Worker $worker, StoreRequest $request)
+    public function update(UpdateRequest $request, Worker $worker)
     {
+
+        $this->authorize('update', $worker);
 
         $data = $request->validated();
         $data['is_married'] = isset($data['is_married']);
+        $data['position_id'] = 1;
 
         $worker->update($data);
 
 
-       return redirect()->route('worker.show', $worker->id);
+       return redirect()->route('workers.show', $worker->id);
 
 
     }
 
-    public function delete(Worker $worker)
+    public function destroy(Worker $worker)
     {
+        $this->authorize('update', $worker);
 
         $worker->delete();
 
-        return redirect()->route('worker.index');
+        return redirect()->route('workers.index');
     }
